fix sx props and decomposition modalWindow and TBThreads
