+++ packages/client/src/components/ModalWindow/index.ts
@@ -1,0 +1,7 @@
+import ModalWindow from './ModalWindow';
+
+export { useModalWindow } from './useModalWindow';
+
+export type { ModalProps } from './types';
+
+export default ModalWindow;
