+++ packages/client/src/components/ModalWindow/types.d.ts
@@ -1,0 +1,6 @@
+export interface ModalProps {
+  open: boolean;
+  handleOpen: () => void;
+  handleClose: () => void;
+  title: ReactNode;
+}
