+++ packages/client/src/features/forum/pages/ThreadList.tsx
@@ -4,9 +4,6 @@ import {
   TableHead,
   TableRow,
   TableCell,
-  TableBody,
-  SxProps,
-  Skeleton,
   Button,
   Box,
 } from '@mui/material';
@@ -20,18 +17,20 @@ import {
 import { useAppDispatch } from '@/app/hooks';
 import { useSelector } from 'react-redux';
 import { chunk } from 'lodash';
-import { formatDateFromUTCString } from '@/shared';
 import { useNavigate, useSearchParams } from 'react-router-dom';
-import { useModal } from '@/components/ModalWindow';
+import { useModalWindow } from '@/components/ModalWindow';
 import { NewThreadModal } from '../components/NewThreadModal';
 import { STATE_STATUSES } from '@/shared/const';
 import {
-  bottomNav,
-  cleanButton,
-  greenButton,
-  mainBox,
-  purpleButton,
+  bottomNavStyles,
+  changePageBtnsStyles,
+  cleanButtonStyles,
+  greenButtonStyles,
+  mainBoxStyles,
+  purpleButtonStyles,
+  tableHeadCellStyles,
 } from '../styles';
+import TBThreads from '../components/TBThreads';
 
 const ForumThreadList: FC = () => {
   const navigate = useNavigate();
@@ -87,24 +86,24 @@ const ForumThreadList: FC = () => {
 
   const isNextPageBtnVisible = chankedThreadList[+currentPagesIdx]?.length > 0;
 
-  const ModalProps = useModal('New Theme');
+  const ModalProps = useModalWindow('New Theme');
 
   return (
     <>
-      <Stack alignItems={'start'} width={'100%'}>
+      <Stack alignItems={'start'} sx={{ width: '100%' }}>
         <BreadCrumbs items={BreadCrumbItems} />
-        <Box sx={{ ...mainBox }}>
+        <Box sx={mainBoxStyles}>
           <Table>
             <TableHead>
               <TableRow>
-                <TableCell sx={tableHeadCellStyles} width={'100%'}>
+                <TableCell sx={{ ...tableHeadCellStyles, width: '100%' }}>
                   Themes
                 </TableCell>
                 <TableCell sx={tableHeadCellStyles}>Messages</TableCell>
                 <TableCell sx={tableHeadCellStyles}>Latest update</TableCell>
               </TableRow>
             </TableHead>
-            <ForumTableBody
+            <TBThreads
               {...{
                 threadList: threadListByPageIdx,
                 threadListStatus,
@@ -112,35 +111,29 @@ const ForumThreadList: FC = () => {
             />
           </Table>
         </Box>
-        <Stack sx={bottomNav}>
-          <Button variant="text" sx={purpleButton} onClick={onBack}>
-            {'<- Back'}
+        <Stack sx={bottomNavStyles}>
+          <Button variant="text" sx={purpleButtonStyles} onClick={onBack}>
+            &lt;- Back
           </Button>
-          <Stack
-            sx={{
-              flexDirection: 'row',
-              flexGrow: '1',
-              justifyContent: 'center',
-              gap: '4rem',
-            }}>
+          <Stack sx={changePageBtnsStyles}>
             <Button
               variant="text"
               disabled={!isPrevPageBtnVisible}
               onClick={onPrevPage}
-              sx={cleanButton}>
-              {'<Prev Page'}
+              sx={cleanButtonStyles}>
+              &lt;Prev Page
             </Button>
             <Button
               variant="text"
               disabled={!isNextPageBtnVisible}
-              sx={cleanButton}
+              sx={cleanButtonStyles}
               onClick={onNextPage}>
-              {'Next Page>'}
+              Next Page&gt;
             </Button>
           </Stack>
           <Button
             variant="text"
-            sx={greenButton}
+            sx={greenButtonStyles}
             onClick={ModalProps.handleOpen}>
             New Theme
           </Button>
@@ -149,112 +142,6 @@ const ForumThreadList: FC = () => {
       <NewThreadModal {...ModalProps} />
     </>
   );
-};
-
-type FTableBodyProps = {
-  threadList?: ForumThreadList;
-  threadListStatus: StateStatus;
-};
-
-// Это нормально или все-таки внести в отдельный файл?
-const ForumTableBody: FC<FTableBodyProps> = ({
-  threadListStatus,
-  threadList,
-}) => {
-  const navigate = useNavigate();
-
-  const toThread = ({ id }: ForumThread) => {
-    navigate(id);
-  };
-
-  const tableRows = new Array(9).fill(null);
-  const tableCells = new Array(3).fill(null);
-  if (threadListStatus === STATE_STATUSES.LOADING) {
-    return (
-      <TableBody>
-        {tableRows.map((notNeed, index) => {
-          return (
-            <TableRow key={index} sx={tableBodyRowStyles}>
-              {tableCells.map((notNeed, index) => (
-                <TableCell sx={tableBodyCellStyles} key={index}>
-                  <Skeleton sx={{ fontSize: '1rem' }} variant="rectangular" />
-                </TableCell>
-              ))}
-            </TableRow>
-          );
-        })}
-      </TableBody>
-    );
-  } else if (threadListStatus === STATE_STATUSES.IDLE) {
-    return (
-      <TableBody>
-        {threadList?.map((thread, index) => (
-          <TableRow
-            key={index}
-            sx={tableBodyRowStyles}
-            onClick={() => toThread(thread)}>
-            <TableCell
-              className={'forum__table-cell'}
-              sx={tableBodyCellStyles}
-              width={'100%'}>
-              {thread.theme}
-            </TableCell>
-            <TableCell
-              className={'forum__table-cell'}
-              sx={{ ...tableBodyCellStyles, textAlign: 'center' }}>
-              {thread.messagesCount}
-            </TableCell>
-            <TableCell
-              className={'forum__table-cell'}
-              sx={{ ...tableBodyCellStyles, textAlign: 'center' }}>
-              {formatDateFromUTCString(thread.lastUpdate)}
-            </TableCell>
-          </TableRow>
-        ))}
-      </TableBody>
-    );
-  } else if (threadListStatus === STATE_STATUSES.FAILED) {
-    return (
-      <TableBody>
-        <TableRow>
-          <TableCell>Something went wrong</TableCell>
-        </TableRow>
-      </TableBody>
-    );
-  }
-  return <></>;
-};
-
-const tableCellStyles: SxProps = {
-  whiteSpace: 'nowrap',
-  padding: '0 0.875rem',
-  borderBottom: '0px',
-  border: '0px',
-};
-
-const tableBodyRowStyles: SxProps = {
-  border: '0px',
-  '& .forum__table-cell': {
-    color: '#fff',
-    transition: '0.1s ease-in color',
-  },
-  '&:hover': {
-    border: '0px',
-    '& .forum__table-cell': {
-      color: 'primary.main',
-      cursor: 'pointer',
-    },
-  },
-};
-
-const tableHeadCellStyles: SxProps = {
-  ...tableCellStyles,
-  color: 'primary.main',
-};
-
-const tableBodyCellStyles: SxProps = {
-  ...tableCellStyles,
-  padding: '0.5rem 1.125rem',
 };
 
 export default ForumThreadList;
