+++ packages/client/src/features/forum/types.d.ts
@@ -1,0 +1,4 @@
+type FTableBodyProps = {
+  threadList?: ForumThreadList;
+  threadListStatus: StateStatus;
+};
